Move temporary type handeling to it's own TU
