+++ src/dawn/IIR/StencilInstantiation.h
@@ -113,19 +113,19 @@ public:
   //                            std::shared_ptr<Expr>& expr, RenameDirection dir);
 
   /// @brief Rename all occurences of field `oldAccessID` to `newAccessID`
-  void renameAllOccurrences(Stencil* stencil, int oldAccessID, int newAccessID);
+  // void renameAllOccurrences(Stencil* stencil, int oldAccessID, int newAccessID);
 
   /// @brief Promote the local variable, given by `AccessID`, to a temporary field
   ///
   /// This will take care of registering the new field (and removing the variable) as well as
   /// replacing the variable accesses with point-wise field accesses.
-  void promoteLocalVariableToTemporaryField(Stencil* stencil, int AccessID,
-                                            const Stencil::Lifetime& lifetime,
-                                            TemporaryScope temporaryScope);
+  // void promoteLocalVariableToTemporaryField(Stencil* stencil, int AccessID,
+  //                                           const Stencil::Lifetime& lifetime,
+  //                                           TemporaryScope temporaryScope);
 
   /// @brief Promote the temporary field, given by `AccessID`, to a real storage which needs to be
   /// allocated by the stencil
-  void promoteTemporaryFieldToAllocatedField(int AccessID);
+  // void promoteTemporaryFieldToAllocatedField(int AccessID);
 
   /// @brief Demote the temporary field, given by `AccessID`, to a local variable
   ///
@@ -134,8 +134,8 @@ public:
   ///
   /// This implicitcly assumes the first access (i.e `lifetime.Begin`) to the field is an
   /// `ExprStmt` and the field is accessed as the LHS of an `AssignmentExpr`.
-  void demoteTemporaryFieldToLocalVariable(Stencil* stencil, int AccessID,
-                                           const Stencil::Lifetime& lifetime);
+  // void demoteTemporaryFieldToLocalVariable(Stencil* stencil, int AccessID,
+  //                                          const Stencil::Lifetime& lifetime);
 
   /// @brief Get the list of stencils
   inline const std::vector<std::unique_ptr<Stencil>>& getStencils() const {
